+++ CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## 3.6.13
+
+### Changed
+
+- Timeline element and details views are now based on interfaces and thus can be replaced
+
 ## 3.6.12
 
 ### Added
@@ -34,10 +40,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Boostrap wizard graphics are not inverted and have visible background
-- Don't block opening room database
 - Remove timeline flickering due to switching between image fallback and thumbnail in reply elements
 - Fix root lifecycle not starting
 - Back Button doesn't work in BlockedContacts view, when focussing a pdf and in the Wizards
+- Add lifecycle stop/start when moving iOS app to background/foreground
 - Export not working in web
 - Don't export media when downloads are disabled
 - Escape CSV export with additional '